requirements.txt

Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
python-dotenv==1.0.0
Werkzeug==2.3.7
openpyxl==3.1.2
reportlab==4.0.4
Flask-Login==0.6.3
alembic==1.12.0
Flask-Migrate==4.0.5
email-validator==2.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
Pillow==10.3.0
Flask-Mail==0.9.1
requests==2.31.0
